requirements.txt
allure_python_commons==2.13.2
pytest==7.4.0
selenium==4.10.0
